Move SupplementaryOutputPaths to Basic.
